feat: improve khaos consistency
