+++ src/core/hooks/authentication/useForgot.ts
@@ -1,16 +1,17 @@
+import Khaos from "@khaos/index";
+import useKhaos from "@khaos/hook";
+import { KhaosHook } from "@khaos/types";
+
 import { ForgotModel } from "@core/models";
 import { ForgotInputs } from "@modules/authentication/forgot/validation";
-import Khaos from "@shared/utilities/Khaos";
-import useKhaos from "@core/hooks/useKhaos";
-import Hook from "@core/interfaces/Hook";
 
-const useForgot = (): Hook<ForgotModel, ForgotInputs> => {
-  const { data, loading, error, fetch } = useKhaos<ForgotModel>();
+const useForgot = (): KhaosHook<ForgotModel, ForgotInputs> => {
+  const { data, loading, error, invoke: fetcher } = useKhaos<ForgotModel>();
   const khaos = new Khaos("/auth/forgot-password").setHttpMethod("POST");
 
   const invoke = async (inputs: ForgotInputs) => {
     khaos.setBody(inputs);
-    await fetch(khaos);
+    await fetcher(khaos);
   };
 
   return { data, loading, error, invoke };
